requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
httpx==0.28.1
pydantic-settings==2.7.0
python-multipart==0.0.20
pytest==8.3.4
openai-whisper==20250625
sounddevice==0.5.2
websockets==15.0.1
